+++ src/entrypoint/rest/http/models/list_objectsv2_response.h
@@ -39,7 +39,7 @@ namespace uh::cluster::rest::http::model
         std::string m_delimiter {};
 
         bool m_maxKeysHasBeenSet = true;
-        int m_maxKeys = 1000;
+        size_t m_maxKeys = 1000;
 
         bool m_commonPrefixesHasBeenSet = false;
         std::vector<std::string> m_commonPrefixes;
